move 'extractLgTemplateRefs' inside 'shared' lib
