+++ Composer/packages/lib/shared/src/lgUtils/index.ts
@@ -4,9 +4,10 @@
 /** models */
 export { default as LgMetaData } from './models/LgMetaData';
 export { default as LgTemplateRef } from './models/LgTemplateRef';
-export { LgTemplateRefPattern } from './parsers/patterns';
 
 /** parsers */
 export { default as parseLgTemplateName } from './parsers/parseLgTemplateName';
 export { default as parseLgTemplateRef } from './parsers/parseLgTemplateRef';
 export { default as parseLgText } from './parsers/parseLgText';
+
+export { default as extractLgTemplateNames } from './extractLgTemplateNames';
